+++ packages/sdk/src/utils/index.ts
@@ -15,3 +15,4 @@ export {
 } from './simulateContract.js'
 export { safeStructuredClone } from './structuredClone.js'
 export { repeatUntilOk } from './repeatUntilOk.js'
+export { prepareHyperliquidSignatureStep } from './hyperliquid.js'