+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @reservoir0x/relay-kit-hooks
 
+## 1.11.3
+
+### Patch Changes
+
+- Updated dependencies [1a88de0]
+  - @reservoir0x/relay-sdk@2.3.2
+
+## 1.11.2
+
+### Patch Changes
+
+- Updated dependencies [d1aeceb]
+  - @reservoir0x/relay-sdk@2.3.1
+
+## 1.11.1
+
+### Patch Changes
+
+- 7985e6b: Add relay trending tokens to token selector
+
 ## 1.11.0
 
 ### Minor Changes